sm8250-common:Declare support for blur
The Adreno 660 should be more than capable of handling this even though
we know that there is a slight performance hit. However, this shouldn't
matter all that much as blur isn't really used in a lot of places.

This also enables a runtime toggle in Developer Options that disables it
in case someone doesn't want it.